+++ Dyna/docs/release-notes.md
@@ -10,9 +10,9 @@
 * Changed the Lock and Load (and by extension Custom Charge) cards.
 * Changed the Burst Charge.
 * Changed the Swift Charge.
+* Changed the Unstable Compound artifact.
 * Fixed Blastwave triggering at the wrong spot for attacks which move the enemy around.
 * Fixed the Bazooka Mod (Dyna-Isaac duo) artifact also triggering on enemy drones.
-* Improved tooltip wording.
 
 ## 1.4.2
 Released 7 March 2025.